requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
black==21.4b2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
cloudpickle==2.2.1
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
dnspython==2.3.0
duckdb==0.7.1
effdet==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
faiss==1.7.4
fastapi==0.95.1
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iopath==0.1.9
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.145
layoutparser==0.3.4
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==2.2.1
portalocker==2.7.0
posthog==3.0.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
streamlit==1.21.0
streamlit-chat==0.0.2.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tesseract==0.1.3
threadpoolctl==3.1.0
tiktoken==0.3.3
timm==0.6.13
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3
tqdm==4.65.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.5.13
unstructured-inference==0.3.2
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
Wand==0.6.11
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.1.0
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
zstandard==0.21.0
